+++ src/main/java/com/example/sms_app/Service/SmsService.java
@@ -53,22 +53,6 @@ public class SmsService {
         return null;
     }
 
-    @RabbitListener(queues = "queue1")
-    public void consumeQueue1(String message) {
-        try {
-            System.out.println("Consumed from Queue 1: " + message);
-        } catch (Exception ex) {
-            throw new CustomException("Error consuming message from Queue 1: " + ex.getMessage());
-        }
-    }
 
-    @RabbitListener(queues = "queue2")
-    public void consumeQueue2(String message) {
-        try {
-            System.out.println("Consumed from Queue 2: " + message);
-        } catch (Exception ex) {
-            throw new CustomException("Error consuming message from Queue 2: " + ex.getMessage());
-        }
-    }
 }
